DomainRow test convert to RTL and add case for Calculated Field
